% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionbuttonUI.R
\name{actionbuttonUI}
\alias{actionbuttonUI}
\title{Create ActionButtons for UI}
\usage{
actionbuttonUI(labelAbtnUI, conditionAbtnUI)
}
\arguments{
\item{labelAbtnUI}{Character. Label for the ActionButtons}

\item{conditionAbtnUI}{Condition for the ActionButton building}
}
\description{
Create ActionButtons for UI
}
