% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_app.R
\name{img_app}
\alias{img_app}
\title{Setup Classification App to Label Images from Local Folder.}
\usage{
img_app(sourceDir, label, multi_label = FALSE, classSize = 100)
}
\arguments{
\item{sourceDir}{The path to the directory where the images to be labeled are located.}

\item{label}{label A `character` `vector` that contains the label for each class.}

\item{multi_label}{A `Boolean` value that indicates if multiple label (TRUE) can be assigned to one image or not (FALSE).
The app launches with check boxes if "multi_label = TRUE" and with action buttons if "multi_label = FALSE".}

\item{classSize}{A `numeric` value with the desired number of images per class (label) for tracking purposes.}
}
\description{
Sets up and launches a Shiny application to manually classify images from a local folder.
}
\note{
he app creates a folder for each label inside "sourceDir" and copies the images from "sourceDir" into the
  corresponding subfolder. Afterwards, the image is removed from "sourceDir".
}
\examples{
\dontrun{
img_app(
  sourceDir = "Dir/to/images",
  label = c("flowering", "fruiting"),
  multi_label = FALSE,
  classSize = 100
)
}
}
