% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_multimedia.R
\name{preprocess_multimedia}
\alias{preprocess_multimedia}
\title{Pre-Process GBIF Multimedia File}
\usage{
preprocess_multimedia(
  multimedia,
  occurrence,
  herbarium.rm = TRUE,
  license.rm = c("all rights reserved", "by-sa", "by-nc", "NA", "unclear")
)
}
\arguments{
\item{multimedia}{A `data.frame` containing the multimedia data or the path to the multimedia file (with the extension "txt", "csv", or "feather").}

\item{occurrence}{A `data.frame` containing the occurrence data or the path to the occurrence file (with the extension "txt", "csv", or "feather").}

\item{herbarium.rm}{A `boolean` indicating whether to remove Herbarium database records (the default is `TRUE`). When set to `FALSE`, Herbarium data is retained.}

\item{license.rm}{A `character vector` specifying the license types to be excluded. Common licenses include 'all rights reserved', 'by-sa', 'by-nc', 'not applicable (NA)', and 'unclear'. By default, 'all rights reserved' and 'unclear' are excluded.}
}
\value{
A `data.frame` containing the pre-processed multimedia data.
}
\description{
Pre-processes a GBIF multimedia file by filtering out records without a GBIF ID (occurrence ID) or image link, processing URLs to images in the iNaturalist database, and optionally removing images from Herbarium databases if the occurrence file contains information about the publisher. Users can also specify which image licenses to exclude from the dataset.
}
\examples{
\dontrun{
# Using default parameters with a file path
preprocess_multimedia(multimedia = "path/to/multimedia.txt", occurrence = "path/to/occurrence.txt")

# Using default parameters with data.frame
preprocess_multimedia(multimedia = multimedia_df, occurrence = occurrence_df)

# Custom parameters, retaining Herbarium data and excluding additional licenses
preprocess_multimedia(
  multimedia = multimedia_df, occurrence = occurrence_df,
  herbarium.rm = FALSE, license.rm = c("all rights reserved", "by-nc", "by-nd")
)
}
}
