% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate files such as images that are stored on disk}
\usage{
remove_duplicates(folder_path, rm_duplicates = TRUE, algorithm = "md5")
}
\arguments{
\item{folder_path}{A `character vector` containing the path to the folder that is to be evaluated.}

\item{rm_duplicates}{A `Boolean` value that indicates if the detected duplicates should be deleted.
`TRUE` by default.}

\item{algorithm}{A `character vector` to select the algorithm that the digest::digest function should use.
"md5" is the default. Other alternatives are "sha1", "crc32", "sha256", "sha512", "xxhash32", "xxhash64", "murmur32", "spookyhash", "blake3", and "crc32c".}
}
\value{
A `character vector` containing the names of the detected (and if rm_duplicates = TRUE, deleted) duplicates.
}
\description{
This function uses the digest::digest function with the md5 algorithm to create hashes with parallel processing using the foreach package.
}
\examples{
\dontrun{
remove_duplicates(folder_path = "path/to/our/folder_with_images", rm_duplicates = TRUE)
}
}
