% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_format.R
\name{remove_format}
\alias{remove_format}
\title{Remove Records with JSON Application Format}
\usage{
remove_format(m.df)
}
\arguments{
\item{m.df}{A `data.frame` containing the multimedia data.}
}
\value{
A `data.frame` without records where the format is a JSON application.
}
\description{
Filters out records from the multimedia data frame that have a format indicating a JSON application.
}
