% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_herbarium.R
\name{remove_herbarium}
\alias{remove_herbarium}
\title{Remove Herbarium Databases from Multimedia Data Frame}
\usage{
remove_herbarium(m.df = multimedia, o.df = occurrence)
}
\arguments{
\item{m.df}{A `data.frame` containing the multimedia data.}

\item{o.df}{A `data.frame` containing the occurrence data.}
}
\value{
A `data.frame` where Herbarium database entries have been removed from the multimedia file.
}
\description{
Removes entries from a multimedia data frame that correspond to Herbarium database records in an occurrence data frame based on the publisher's information.
}
\note{
This function assumes that the `publisher` column in `o.df` contains Herbarium information. This is not always the case.
}
