% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_license.R
\name{remove_license}
\alias{remove_license}
\title{Remove gbifIDs with Specified Licenses from Multimedia Data Frame}
\usage{
remove_license(m.df = multimedia.df, license = license.rm)
}
\arguments{
\item{m.df}{A `data.frame` containing multimedia data.}

\item{license}{A `character vector` specifying the license types to be excluded from further processing.
Accepted values include 'all rights reserved', 'by-sa', 'by-nc', 'NA', and 'unclear'.}
}
\value{
A `data.frame` with records filtered according to the specified license criteria.
}
\description{
Filters out records from a multimedia data frame based on specified license types.
By default, records with "all rights reserved" and "unclear" licenses are removed.
}
