% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_static.inaturalist.R
\name{replace_static.inaturalist}
\alias{replace_static.inaturalist}
\title{Replace Static iNaturalist URLs}
\usage{
replace_static.inaturalist(m.df)
}
\arguments{
\item{m.df}{A `data.frame` containing the multimedia data.}
}
\value{
A `data.frame` with the updated URLs in the identifier column.
}
\description{
Replaces static iNaturalist URLs in the identifier column of a multimedia `data frame` with the correct URL format.
}
