% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.write_image.R
\name{server.write_image}
\alias{server.write_image}
\title{Write Images to Disk (Server Helper Function)}
\usage{
server.write_image(
  multi_label,
  image_cache,
  image_format,
  currentRows,
  imageSelection,
  multimedia,
  backupDir
)
}
\description{
The `server.write_image` function saves images to a specified directory. It handles both single-label and multi-label cases.
 Images are saved in a user-specified format (e.g., PNG, JPG) and are organized into directories based on their labels.
}
