% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_app.R
\name{setup_app}
\alias{setup_app}
\title{Setup Classification App to Label Images from GBIF Multimedia File}
\usage{
setup_app(
  multimedia,
  classSize = 100,
  label,
  multi_label = FALSE,
  backupDir,
  backupInterval = 1
)
}
\arguments{
\item{multimedia}{A `data.frame` or the path to the GBIF multimedia file (with the extension "txt", "csv", or "feather").}

\item{classSize}{A `numeric` value with the desired number of images per class (label) for tracking purposes.}

\item{label}{A `character` `vector` that contains the label for each class.}

\item{multi_label}{A `Boolean` value that indicates if multiple label (TRUE) can be assigned to one image or not (FALSE).
The app launches with check boxes if "multi_label = TRUE" and with action buttons if "multi_label = FALSE".}

\item{backupDir}{A `character` string that specifies the path to the directory for storing backups and the final file.}

\item{backupInterval}{A `numeric` value that indicates the interval for automatic backups of the multimedia file (in minutes).}
}
\value{
A `shiny` app for image classification is launched, and a modified `data.frame` is written into "backupDir".
}
\description{
Sets up and launches a Shiny application to manually classify images from a GBIF multimedia file.
}
\note{
The modified `data.frame` is written using the `feather` package. The assigned labels are stored in a new "label" column as character vectors.
}
\examples{
\dontrun{
setup_app(
  multimedia = "path/to/multimedia.txt",
  classSize = 100,
  label = c("flowering", "fruiting"),
  multi_label = FALSE,
  backupDir = "path/to/backup",
  backupInterval = 1
)
}
}
