% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try2read.R
\name{try2read}
\alias{try2read}
\title{try to read the images from URL in multimedia file}
\usage{
try2read(identifier)
}
\arguments{
\item{identifier}{The image URL.}
}
\description{
try to read the images from URL in multimedia file
}
