% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_date.R
\name{remove_date}
\alias{remove_date}
\title{Remove Dates without Time Specification and Dates recorded on 01.01. of a year at 00.00.00.}
\usage{
remove_date(m.df = multimedia, date.rm = TRUE, time.rm = FALSE)
}
\arguments{
\item{m.df}{A `data.frame` containing the multimedia data.}

\item{rm_wo_times}{A `boolean` value to remove dates without time specifications. If `TRUE`, the dates will be removed. The default is `FALSE`.}
}
\description{
The function removes images without a specified date of creation (if date.rm = TRUE)
and removes images without a specified time and suspicious date and time combinations, i. e. 01.01.yyyy 00:00:00.
}
